+++ workspace/solutions/kotlin/lib/src/test/kotlin/s0/s2/s3/s9/s_0239_test.kt
@@ -1,0 +1,16 @@
+package s0.s2.s3.s9
+
+import utils.T
+import kotlin.test.Test
+import s0.s2.s3.s9.s_0239_00.Solution as s_0239_00
+
+class s_0239_test {
+    @Test
+    fun test() {
+        val t = T(
+            this::class,
+            s_0239_00::class,
+        )
+        t.run()
+    }
+}