+++ workspace/solutions/rust/src/s0/mod.rs
@@ -1,5 +1,6 @@
 pub mod s0;
 pub mod s1;
+pub mod s2;
 pub mod s3;
 pub mod s4;
 pub mod s7;