+++ workspace/solutions/rust/src/s0/s2/s3/s9/mod.rs
@@ -1,0 +1,4 @@
+pub mod s_0239_00;
+pub mod s_0239_test;
+
+pub type TSolution = fn(Vec<i32>, i32) -> Vec<i32>;